% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_indexages.R
\name{gen_indexdist}
\alias{gen_indexdist}
\title{get index date for cases}
\usage{
gen_indexdist(
  dbf,
  idxtab = "fullindexages",
  pats = "patients",
  ctype = "\%",
  seed = NULL
)
}
\description{
get index date for cases
}
