% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohorts.R
\name{get_cohort}
\alias{get_cohort}
\title{get the cohorts}
\usage{
get_cohort(
  dbf,
  codetable = "cohortcodes",
  codetype = "\%",
  minage = 0,
  maxage = 16,
  indexfield = "case",
  vfill = 0
)
}
\description{
get the cohorts
}
