% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_drugissues.R
\name{load_drugissues}
\alias{load_drugissues}
\title{get the drugissue records and convert some fields to useful field types}
\usage{
load_drugissues(
  pddir,
  dbf,
  ow = F,
  db = F,
  tab_name = "drug_issues",
  add = F,
  selvars2 = c("patid", "prodcodeid", "issuedate", "dosageid", "quantity", "quantunitid",
    "duration")
)
}
\arguments{
\item{batch}{the name of the batch subdirectory}

\item{datadir}{the name of the data directory}

\item{bpp}{BiocParallal Multicore Parameters}
}
\description{
get the drugissue records and convert some fields to useful field types
}
