% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hesfiles.R
\name{load_hesfiles}
\alias{load_hesfiles}
\title{load hes files}
\usage{
load_hesfiles(pddir, dbf, ow = F, db = F, tad, pats)
}
\arguments{
\item{pddir}{directory containing hes data}

\item{dbf}{name of duckdb database}

\item{ow}{overwrite an existing file}

\item{db}{debug switch}

\item{tad}{text adjustment to remove from file names when creating table name}

\item{pats}{a list of patient ids to extract from the hes data}
}
\description{
get the hes files records and convert some fields to useful field types
}
