% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obs.R
\name{load_obs}
\alias{load_obs}
\title{get the drugissue records and convert some fields to useful field types}
\usage{
load_obs(
  pddir,
  dbf,
  ow = F,
  db = F,
  tab_name = "observations",
  selvars = c("patid", "consid", "parentobsid", "probobsid", "medcodeid", "obsdate",
    "value", "numunitid", "numrangelow", "numrangehigh")
)
}
\arguments{
\item{bdir}{the name of the batch subdirectory}

\item{odir}{the name of the output directory}

\item{olist}{the list of observation codes}

\item{bpp}{BiocParallal Multicore Parameters


Pass a table of covariate codes and generate covariates table}
}
\description{
get the drugissue records and convert some fields to useful field types
}
