% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_table.R
\name{load_table}
\alias{load_table}
\title{get the drugissue records and convert some fields to useful field types}
\usage{
load_table(
  filename,
  dbf,
  ow = F,
  db = F,
  tab_name = gsub("(^[0-9]*_|[.].*)", "", basename(filename)),
  selvars = NULL,
  delim = "\\t"
)
}
\arguments{
\item{bdir}{the name of the batch subdirectory}

\item{odir}{the name of the output directory}

\item{olist}{the list of observation codes}

\item{bpp}{BiocParallal Multicore Parameters


Pass a table of covariate codes and generate covariates table}
}
\description{
get the drugissue records and convert some fields to useful field types
}
