% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_patients.R
\name{proc_patients}
\alias{proc_patients}
\title{load patients}
\usage{
proc_patients(
  dbpost = dbpost,
  dbpre = dbpre,
  ow = F,
  db = F,
  newtab = "patients",
  oldtab = "patients",
  jointabs = list(opats = "death_patient", hpats = "hes_patient", ipats =
    "patient_2019_imd")
)
}
\arguments{
\item{batch}{the name of the batch subdirectory}

\item{datadir}{the name of the data directory}
}
\description{
get the patient records and convert some fields to useful field types
}
