% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_test.R
\name{run_test}
\alias{run_test}
\title{run test}
\usage{
run_test(
  dbf,
  cohort = NULL,
  symptomtab = NULL,
  cortype = "bon",
  ratioval = 1,
  pcut = 0.05,
  groupat = "medcodeid"
)
}
\description{
run test
}
