% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_test2.R
\name{run_test2}
\alias{run_test2}
\title{run test}
\usage{
run_test2(
  dbf,
  cohort = NULL,
  symptomtab = NULL,
  cortype = "bon",
  ratioval = 1,
  pcut = 0.05,
  groupat = "medcodeid"
)
}
\description{
run test
}
