% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cons.R
\name{load_cons}
\alias{load_cons}
\title{get observations batch}
\usage{
load_cons(
  pddir,
  dbf,
  ow = F,
  db = F,
  tab_name = "consultations",
  add = F,
  selvars = c("patid", "consid", "consdate", "conssourceid", "cprdconstype",
    "consmedcodeid")
)
}
\arguments{
\item{bdir}{the name of the batch subdirectory}

\item{odir}{the name of the output directory}

\item{olist}{the list of observation codes}

\item{bpp}{BiocParallal Multicore Parameters


Pass a table of covariate codes and generate covariates table}
}
\description{
get the drugissue records and convert some fields to useful field types
}
