% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hesfiles.R
\name{load_hesfiles}
\alias{load_hesfiles}
\title{get the drugissue records and convert some fields to useful field types}
\usage{
load_hesfiles(pddir, dbf, ow = T, db = F, tad, pats)
}
\arguments{
\item{bdir}{the name of the batch subdirectory}

\item{odir}{the name of the output directory}

\item{olist}{the list of observation codes}

\item{bpp}{BiocParallal Multicore Parameters


Pass a table of covariate codes and generate covariates table}
}
\description{
get the drugissue records and convert some fields to useful field types
}
